+++ agrobr_qgis/core/exceptions.py
@@ -1,0 +1,26 @@
+class AgroBRError(Exception):
+    pass
+
+
+class FetchError(AgroBRError):
+    pass
+
+
+class ContractError(AgroBRError):
+    pass
+
+
+class JoinError(AgroBRError):
+    pass
+
+
+class AuthError(AgroBRError):
+    pass
+
+
+class DependencyError(AgroBRError):
+    pass
+
+
+class ChecksumError(AgroBRError):
+    pass